Remove support for MariaDB 10.4 and MySQL 5.7
