+++ src/Platforms/Keywords/MySQLKeywords.php
@@ -19,10 +19,12 @@ class MySQLKeywords extends KeywordList
         return [
             'ACCESSIBLE',
             'ADD',
+            'ADMIN',
             'ALL',
             'ALTER',
             'ANALYZE',
             'AND',
+            'ARRAY',
             'AS',
             'ASC',
             'ASENSITIVE',
@@ -48,6 +50,8 @@ class MySQLKeywords extends KeywordList
             'CONVERT',
             'CREATE',
             'CROSS',
+            'CUBE',
+            'CUME_DIST',
             'CURRENT_DATE',
             'CURRENT_TIME',
             'CURRENT_TIMESTAMP',
@@ -65,6 +69,7 @@ class MySQLKeywords extends KeywordList
             'DEFAULT',
             'DELAYED',
             'DELETE',
+            'DENSE_RANK',
             'DESC',
             'DESCRIBE',
             'DETERMINISTIC',
@@ -77,13 +82,16 @@ class MySQLKeywords extends KeywordList
             'EACH',
             'ELSE',
             'ELSEIF',
+            'EMPTY',
             'ENCLOSED',
             'ESCAPED',
             'EXISTS',
+            'EXCEPT',
             'EXIT',
             'EXPLAIN',
             'FALSE',
             'FETCH',
+            'FIRST_VALUE',
             'FLOAT',
             'FLOAT4',
             'FLOAT8',
@@ -92,10 +100,13 @@ class MySQLKeywords extends KeywordList
             'FOREIGN',
             'FROM',
             'FULLTEXT',
+            'FUNCTION',
             'GENERATED',
             'GET',
             'GRANT',
             'GROUP',
+            'GROUPING',
+            'GROUPS',
             'HAVING',
             'HIGH_PRIORITY',
             'HOUR_MICROSECOND',
@@ -123,10 +134,15 @@ class MySQLKeywords extends KeywordList
             'IO_BEFORE_GTIDS',
             'IS',
             'ITERATE',
+            'JSON_TABLE',
             'JOIN',
             'KEY',
             'KEYS',
             'KILL',
+            'LAG',
+            'LAST_VALUE',
+            'LATERAL',
+            'LEAD',
             'LEADING',
             'LEAVE',
             'LEFT',
@@ -150,6 +166,7 @@ class MySQLKeywords extends KeywordList
             'MEDIUMBLOB',
             'MEDIUMINT',
             'MEDIUMTEXT',
+            'MEMBER',
             'MIDDLEINT',
             'MINUTE_MICROSECOND',
             'MINUTE_SECOND',
@@ -158,8 +175,11 @@ class MySQLKeywords extends KeywordList
             'NATURAL',
             'NO_WRITE_TO_BINLOG',
             'NOT',
+            'NTH_VALUE',
+            'NTILE',
             'NULL',
             'NUMERIC',
+            'OF',
             'ON',
             'OPTIMIZE',
             'OPTIMIZER_COSTS',
@@ -170,16 +190,22 @@ class MySQLKeywords extends KeywordList
             'OUT',
             'OUTER',
             'OUTFILE',
+            'OVER',
             'PARTITION',
+            'PERCENT_RANK',
+            'PERSIST',
+            'PERSIST_ONLY',
             'PRECISION',
             'PRIMARY',
             'PROCEDURE',
             'PURGE',
+            'RANK',
             'RANGE',
             'READ',
             'READ_WRITE',
             'READS',
             'REAL',
+            'RECURSIVE',
             'REFERENCES',
             'REGEXP',
             'RELEASE',
@@ -193,6 +219,9 @@ class MySQLKeywords extends KeywordList
             'REVOKE',
             'RIGHT',
             'RLIKE',
+            'ROW',
+            'ROWS',
+            'ROW_NUMBER',
             'SCHEMA',
             'SCHEMAS',
             'SECOND_MICROSECOND',
@@ -216,6 +245,7 @@ class MySQLKeywords extends KeywordList
             'STARTING',
             'STORED',
             'STRAIGHT_JOIN',
+            'SYSTEM',
             'TABLE',
             'TERMINATED',
             'THEN',
@@ -247,6 +277,7 @@ class MySQLKeywords extends KeywordList
             'WHEN',
             'WHERE',
             'WHILE',
+            'WINDOW',
             'WITH',
             'WRITE',
             'XOR',
