--- src/Platforms/MySQL80Platform.php
@@ -1,27 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Doctrine\DBAL\Platforms;
-
-use Doctrine\DBAL\Platforms\Keywords\KeywordList;
-use Doctrine\DBAL\Platforms\Keywords\MySQL80Keywords;
-use Doctrine\DBAL\SQL\Builder\SelectSQLBuilder;
-
-/**
- * Provides the behavior, features and SQL dialect of the MySQL 8.0 (8.0 GA) database platform.
- *
- * @deprecated This class will be removed once support for MySQL 5.7 is dropped.
- */
-class MySQL80Platform extends MySQLPlatform
-{
-    protected function createReservedKeywordsList(): KeywordList
-    {
-        return new MySQL80Keywords();
-    }
-
-    public function createSelectSQLBuilder(): SelectSQLBuilder
-    {
-        return AbstractPlatform::createSelectSQLBuilder();
-    }
-}
