+++ tests/Driver/VersionAwarePlatformDriverTest.php
@@ -7,10 +7,8 @@ namespace Doctrine\DBAL\Tests\Driver;
 use Doctrine\DBAL\Connection\StaticServerVersionProvider;
 use Doctrine\DBAL\Driver;
 use Doctrine\DBAL\Platforms\AbstractPlatform;
-use Doctrine\DBAL\Platforms\MariaDB1052Platform;
 use Doctrine\DBAL\Platforms\MariaDB1060Platform;
 use Doctrine\DBAL\Platforms\MariaDBPlatform;
-use Doctrine\DBAL\Platforms\MySQL80Platform;
 use Doctrine\DBAL\Platforms\MySQLPlatform;
 use Doctrine\DBAL\Platforms\PostgreSQLPlatform;
 use PHPUnit\Framework\Attributes\DataProvider;
@@ -34,13 +32,11 @@ class VersionAwarePlatformDriverTest extends TestCase
     public static function mySQLVersionProvider(): array
     {
         return [
-            ['5.7.0', MySQLPlatform::class],
-            ['8.0.11', MySQL80Platform::class],
+            ['8.0.11', MySQLPlatform::class],
             ['5.5.40-MariaDB-1~wheezy', MariaDBPlatform::class],
             ['5.5.5-MariaDB-10.2.8+maria~xenial-log', MariaDBPlatform::class],
             ['10.2.8-MariaDB-10.2.8+maria~xenial-log', MariaDBPlatform::class],
-            ['10.2.8-MariaDB-1~lenny-log', MariaDBPlatform::class],
-            ['10.5.2-MariaDB-1~lenny-log', MariaDB1052Platform::class],
+            ['10.5.2-MariaDB-1~lenny-log', MariaDBPlatform::class],
             ['10.6.0-MariaDB-1~lenny-log', MariaDB1060Platform::class],
             ['11.0.2-MariaDB-1:11.0.2+maria~ubu2204', MariaDB1060Platform::class],
         ];
