+++ tests/Functional/Schema/MySQL/ComparatorTest.php
@@ -7,8 +7,6 @@ namespace Doctrine\DBAL\Tests\Functional\Schema\MySQL;
 use Doctrine\DBAL\Exception;
 use Doctrine\DBAL\Platforms\AbstractMySQLPlatform;
 use Doctrine\DBAL\Platforms\AbstractPlatform;
-use Doctrine\DBAL\Platforms\MariaDBPlatform;
-use Doctrine\DBAL\Platforms\MySQL80Platform;
 use Doctrine\DBAL\Schema\AbstractSchemaManager;
 use Doctrine\DBAL\Schema\Column;
 use Doctrine\DBAL\Schema\Comparator;
@@ -198,7 +196,7 @@ final class ComparatorTest extends FunctionalTestCase
 
     public function testMariaDb1043NativeJsonUpgradeDetected(): void
     {
-        if (! $this->platform instanceof MariaDBPlatform && ! $this->platform instanceof MySQL80Platform) {
+        if (! $this->platform instanceof AbstractMySQLPlatform) {
             self::markTestSkipped();
         }
